meant to use floats

+++ LuckParser/Controllers/Parser.cs
@@ -639,17 +639,17 @@ namespace LuckParser.Controllers
                 List<CombatItem> c_Pos = combat_data.getStates(p.getInstid(), ParseEnum.StateChange.Position, boss_data.getFirstAware(), boss_data.getLastAware());
                 foreach (CombatItem c in c_Pos)
                 {
-                    long X = (long)(c.getDstAgent() >> 32);
-                    long Y = (long)c.getDstAgent();
-                    p.AddPositionPoint(c.getTime(), new Point3D(X,Y, c.getValue()));
+                    float X = (float)(c.getDstAgent() >> 32);
+                    float Y = (float)c.getDstAgent();
+                    p.AddPositionPoint(c.getTime(), new Point3D(X,Y, (float)c.getValue()));
                 }
                 //set player velocity
                 List<CombatItem> c_Vel = combat_data.getStates(p.getInstid(), ParseEnum.StateChange.Velocity, boss_data.getFirstAware(), boss_data.getLastAware());
                 foreach (CombatItem c in c_Vel)
                 {
-                    long X = (long)(c.getDstAgent() >> 32);
-                    long Y = (long)c.getDstAgent();
-                    p.AddVelocityPoint(c.getTime(), new Point3D(X,Y, c.getValue()));
+                    float X = (float)(c.getDstAgent() >> 32);
+                    float Y = (float)c.getDstAgent();
+                    p.AddVelocityPoint(c.getTime(), new Point3D(X,Y, (float)c.getValue()));
                 }
             }
